LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/fetch/metadata/generated/element-frame.sub-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY&mime=text%2Fhtml
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY&mime=text%2Fhtml
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY&mime=text%2Fhtml
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY&mime=text%2Fhtml

FAIL sec-fetch-site - Not sent to non-trustworthy same-origin destination assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - Not sent to non-trustworthy same-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-site - Not sent to non-trustworthy cross-site destination assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-mode - Not sent to non-trustworthy same-origin destination assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-mode - Not sent to non-trustworthy same-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-mode - Not sent to non-trustworthy cross-site destination assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-dest - Not sent to non-trustworthy same-origin destination assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
FAIL sec-fetch-dest - Not sent to non-trustworthy same-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-dest - Not sent to non-trustworthy cross-site destination assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
PASS sec-fetch-user - Not sent to non-trustworthy same-origin destination
FAIL sec-fetch-user - Not sent to non-trustworthy same-site destination promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
PASS sec-fetch-user - Not sent to non-trustworthy cross-site destination
FAIL sec-fetch-site - HTTPS downgrade (header not sent) assert_not_own_property: unexpected property "sec-fetch-site" is found on object
PASS sec-fetch-site - HTTPS upgrade
PASS sec-fetch-site - HTTPS downgrade-upgrade
